LibWeb: Define GLintptr as 64-bit
In the IDL file GLintptr is defined as long long which is 64-bit, which
made that definition incorrect.

+++ Libraries/LibWeb/WebGL/Types.h
@@ -15,7 +15,7 @@ using GLenum = unsigned int;
 using GLuint = unsigned int;
 using GLint = int;
 using GLsizei = int;
-using GLintptr = int;
+using GLintptr = long long;
 
 // FIXME: This should really be "struct __GLsync*", but the linker doesn't recognise it.
 //        Since this conflicts with the original definition of GLsync, the suffix "Internal" has been added.
